✨ v2.2.0 New scheduler, timestamps in logs, dashboard improvements, + more. (#154)
# v2.2.0

**New Features**

- **New job scheduler**  
  The job scheduler has been rewritten with a [simpler syntax](https://meerschaum.io/reference/background-jobs/#-schedules).

  ```bash
  mrsm sync pipes -s 'daily & mon-fri starting 00:00 tomorrow' -d
  ```

- **Add `show schedule`.**  
  Validate your schedules' upcoming timestamps with `show schedule`.

  ```
  mrsm show schedule 'daily & mon-fri starting 2024-05-01'

  Next 5 timestamps for schedule 'daily & mon-fri starting 2024-05-01':

    2024-05-01 00:00:00+00:00
    2024-05-02 00:00:00+00:00
    2024-05-03 00:00:00+00:00
    2024-05-06 00:00:00+00:00
    2024-05-07 00:00:00+00:00
  ```

- **Added timestamps to log file lines.**  
  Log files now prepend the current minute to each line of the file, and the timestamps are also printed when viewing logs with `show logs`.
  To disable this behavio, set `MRSM{jobs:logs:timestamps:enabled}` to `false`.

  You may change the timestamp format under the config keys `MRSM{jobs:logs:timestamps:format}` (timestamp written to disk) and `MRSM{jobs:logs:timestamps:follow_format}` (timestamp printed when following via `show logs`.).

- **Add `--skip-deps`.**  
  When installing plugins, you may skip dependencies with `--skip-deps`. This should improve the iteration loop during development.

  ```bash
  mrsm install plugin noaa --no-deps
  ```

- **Add logs buttons to job cards on the Web UI.**  
  For your convenience, "Follow logs" and "Download logs" buttons have been added to jobs' cards.

- **Add a Delete button to job cards on the Web UI.**  
  You may now delete a job from its card (once stopped, that is).

- **Add management buttons to pipes' cards.**  
  For your convenience, you may now sync, verify, clear, drop, and delete pipes directly from cards.

- **Designate your packages as plugins with the `meerschaum.plugins` entry point.**  
  You may now specify your existing packages as Meerschaum plugins by adding the `meerschaum.plugins` [Entrypoint](https://packaging.python.org/en/latest/guides/creating-and-discovering-plugins/#using-package-metadata) to your package metadata:

  ```python
  from setuptools import setup
  
  setup(
      ...,
      entry_points = {
          'meerschaum.plugins': [
              'foo = foo',
          ],
      },
  )
  ```

  or if you are using `pyproject.toml`:

  ```toml
  [project.entry-points."meerschaum.plugins"]
  foo = "foo"
  ```

- **Pre- and post-sync hooks are printed separately.**  
  The results of sync hooks are now printed right after execution rather than after the sync.

**Bugfixes**

- **Fixed a filtering bug on the Web UI when changing instances.**  
  When changing instances on the Web Console, the connector, metric, and location choices will reset appropriately.

- **Ctrl+C when exiting `show logs`.**  
  Pressing Ctrl+C will now exit the `show logs` immediately.

**Breaking Changes**

- **No longer supporting the old scheduler syntax.**  
  If you have jobs with the old scheduler syntax (e.g. using the keyword `before`), you may need to delete and recreate your jobs with an updated schedule.

- **Upgraded to `psycopg` from `psycopg2`.**  
  The upgrade to `psycopg` (version 3) should provide better performance for larger transactions.

- **`Daemon.cleanup()` now returns a `SuccessTuple`.**

**Other changes**

- **Bumped `xterm.js` to v5.5.0.**
- **Added tags to the pipes card.**
- **Replaced `watchgod` with `watchfiles`.**
- **Replaced `rocketry` with `APScheduler`.**
- **Removed `pydantic` from dependencies.**
- **Removed `passlib` from dependencies.**
- **Bumped default TimescaleDB image to `latest-pg16-oss`.**
- **Held back `duckdb` to `<0.10.3`.**

+++ meerschaum/api/dash/callbacks/dashboard.py
@@ -440,17 +440,12 @@ def update_flags(input_flags_dropdown_values, n_clicks, input_flags_texts):
             className = 'input-text',
         )
 
+    remove_index = trigger_dict['index'] if trigger_type == 'input-flags-remove-button' else None
     rows = [
         build_row(i, val, val_text)
         for i, (val, val_text) in enumerate(zip(input_flags_dropdown_values, input_flags_texts))
+        if i != remove_index
     ]
-
-    if trigger_type == 'input-flags-remove-button':
-        remove_index = trigger_dict['index']
-        try:
-            del rows[remove_index]
-        except IndexError:
-            pass
 
     if not rows or input_flags_dropdown_values[-1]:
         rows.append(build_row(len(rows), None, None))
@@ -488,11 +483,11 @@ def update_keys_options(
     """
     ctx = dash.callback_context
     trigger = ctx.triggered[0]['prop_id'].split('.')[0]
+    instance_click = trigger == 'instance-select'
 
     ### Update the instance first.
     update_instance_keys = False
     if not instance_keys:
-        #  instance_keys = get_config('meerschaum', 'web_instance')
         instance_keys = str(get_api_connector())
         update_instance_keys = True
     instance_alerts = []
@@ -516,20 +511,23 @@ def update_keys_options(
     if location_keys:
         num_filter += 1
 
-    _ck_alone, _mk_alone, _lk_alone = False, False, False
-    _ck_filter, _mk_filter, _lk_filter = connector_keys, metric_keys, location_keys
-
-    _ck_alone = connector_keys and num_filter == 1
-    _mk_alone = metric_keys and num_filter == 1
-    _lk_alone = location_keys and num_filter == 1
+    _ck_filter = connector_keys
+    _mk_filter = metric_keys
+    _lk_filter = location_keys
+    _ck_alone = (connector_keys and num_filter == 1) or instance_click
+    _mk_alone = (metric_keys and num_filter == 1) or instance_click
+    _lk_alone = (location_keys and num_filter == 1) or instance_click
 
     from meerschaum.utils import fetch_pipes_keys
 
     try:
         _all_keys = fetch_pipes_keys('registered', get_web_connector(ctx.states))
         _keys = fetch_pipes_keys(
-            'registered', get_web_connector(ctx.states),
-            connector_keys=_ck_filter, metric_keys=_mk_filter, location_keys=_lk_filter
+            'registered',
+            get_web_connector(ctx.states),
+            connector_keys = _ck_filter,
+            metric_keys = _mk_filter,
+            location_keys = _lk_filter,
         )
     except Exception as e:
         instance_alerts += [alert_from_success_tuple((False, str(e)))]
@@ -545,15 +543,39 @@ def update_keys_options(
             k = locals()[key_type]
             if k not in _seen_keys[key_type]:
                 _k = 'None' if k in (None, '[None]', 'None', 'null') else k
-                options.append({'label' : _k, 'value' : _k})
+                options.append({'label': _k, 'value': _k})
                 _seen_keys[key_type].add(k)
 
     add_options(_connectors_options, _all_keys if _ck_alone else _keys, 'ck')
     add_options(_metrics_options, _all_keys if _mk_alone else _keys, 'mk')
     add_options(_locations_options, _all_keys if _lk_alone else _keys, 'lk')
-    connector_keys = [ck for ck in connector_keys if ck in [_ck['value'] for _ck in _connectors_options]]
-    metric_keys = [mk for mk in metric_keys if mk in [_mk['value'] for _mk in _metrics_options]]
-    location_keys = [lk for lk in location_keys if lk in [_lk['value'] for _lk in _locations_options]]
+    _connectors_options.sort(key=lambda x: str(x).lower())
+    _metrics_options.sort(key=lambda x: str(x).lower())
+    _locations_options.sort(key=lambda x: str(x).lower())
+    connector_keys = [
+        ck
+        for ck in connector_keys
+        if ck in [
+            _ck['value']
+            for _ck in _connectors_options
+        ]
+    ]
+    metric_keys = [
+        mk
+        for mk in metric_keys
+        if mk in [
+            _mk['value']
+            for _mk in _metrics_options
+        ]
+    ]
+    location_keys = [
+        lk
+        for lk in location_keys
+        if lk in [
+            _lk['value']
+            for _lk in _locations_options
+        ]
+    ]
     _connectors_datalist = [html.Option(value=o['value']) for o in _connectors_options]
     _metrics_datalist = [html.Option(value=o['value']) for o in _metrics_options]
     _locations_datalist = [html.Option(value=o['value']) for o in _locations_options]
@@ -680,6 +702,9 @@ dash_app.clientside_callback(
     Input({'type': 'pipe-download-csv-button', 'index': ALL}, 'n_clicks'),
 )
 def download_pipe_csv(n_clicks):
+    """
+    Download the most recent chunk as a CSV file.
+    """
     if not n_clicks:
         raise PreventUpdate
     ctx = dash.callback_context.triggered
@@ -688,11 +713,11 @@ def download_pipe_csv(n_clicks):
     pipe = pipe_from_ctx(ctx, 'n_clicks')
     if pipe is None:
         raise PreventUpdate
-    filename = str(pipe.target) + '.csv'
     bounds = pipe.get_chunk_bounds(bounded=True, debug=debug)
-    begin, _ = bounds[-1]
+    begin, end = bounds[-1]
+    filename = str(pipe.target) + f" {begin} - {end}.csv"
     try:
-        df = pipe.get_data(begin=begin, end=None, debug=debug)
+        df = pipe.get_data(begin=begin, end=end, debug=debug)
     except Exception as e:
         df = None
     if df is not None:
@@ -817,6 +842,51 @@ def sync_documents_click(n_clicks, sync_editor_text):
 
     return alert_from_success_tuple((success, msg))
 
+
+dash_app.clientside_callback(
+    """
+    function(n_clicks_arr, url){
+        display_block = {"display": "block"};
+
+        var clicked = false;
+        for (var i = 0; i < n_clicks_arr.length; i++){
+            if (n_clicks_arr[i]){
+                clicked = true;
+                break;
+            }
+        }
+        if (!clicked){ return dash_clientside.no_update; }
+
+        const triggered_id = dash_clientside.callback_context.triggered_id;
+        const action = triggered_id["action"];
+        const pipe_meta = JSON.parse(triggered_id["index"]);
+
+        iframe = document.getElementById('webterm-iframe');
+        if (!iframe){ return dash_clientside.no_update; }
+        var location = pipe_meta.location;
+        if (!pipe_meta.location){
+            location = "None";
+        }
+
+        iframe.contentWindow.postMessage(
+            {
+                action: action,
+                subaction: "pipes",
+                connector_keys: [pipe_meta.connector],
+                metric_keys: [pipe_meta.metric],
+                location_keys: [location],
+                instance: pipe_meta.instance,
+            },
+            url
+        );
+        dash_clientside.set_props("webterm-div", {"style": display_block});
+        return [];
+    }
+    """,
+    Output('content-div-right', 'children'),
+    Input({'type': 'manage-pipe-button', 'index': ALL, 'action': ALL}, 'n_clicks'),
+    State('location', 'href'),
+)
 
 @dash_app.callback(
     Output("navbar-collapse", "is_open"),
